main 페이지 js 통합
