--- src/pages/main/js/contentsSwiper.js
@@ -1,39 +1,0 @@
-/* ----------------- 컨텐츠 스와이퍼 ----------------- */
-
-document.addEventListener('DOMContentLoaded', function () {
-  const swiperConfigs = [
-    { selector: '.swiperMust' },
-    { selector: '.swiperQuick' },
-    { selector: '.swiperPopular' },
-    { selector: '.swiperLive' },
-    { selector: '.swiperOnly' },
-    { selector: '.swiperPromotion' },
-    { selector: '.swiperEvent' },
-  ];
-
-  function createSwipers(configs) {
-    return configs.map(
-      (config) =>
-        new Swiper(config.selector, {
-          // 공통 설정
-          slidesPerView: 'auto',
-          spaceBetween: 8,
-          pagination: {
-            el: `${config.selector} .swiper-pagination`,
-            clickable: true,
-          },
-          navigation: {
-            nextEl: `${config.selector} .swiper-button-next`,
-            prevEl: `${config.selector} .swiper-button-prev`,
-          },
-          breakpoints: {
-            1025: {
-              spaceBetween: 16,
-            },
-          },
-        })
-    );
-  }
-
-  const swipers = createSwipers(swiperConfigs);
-});
